# openapi.yaml
openapi: 3.0.0
info:
  title: Task Management Platform API
  version: 1.0.0
  description: Comprehensive API for task and project management
  contact:
    name: Development Team
    email: dev@taskboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.taskboard.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /tasks/{id}:
    get:
      summary: Get task by ID
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update task
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete task
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/assign:
    put:
      summary: Assign task to user
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTaskRequest'
      responses:
        '200':
          description: Task assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
  /tasks/{id}/status:
    put:
      summary: Change task status
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      responses:
        '200':
          description: Status changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
  /tasks/search:
    get:
      summary: Search tasks
      tags: [Tasks]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'

  # Time Tracking endpoints
...
  /search:
    get:
      summary: Global search
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: types
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [task, project, user, comment]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'

  # Notification endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, IN_REVIEW, DONE, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        projectId:
          type: string
          format: uuid
        assigneeId:
          type: string
          format: uuid
        creatorId:
          type: string
          format: uuid
        labels:
          type: array
          items:
            type: string
        timeSpent:
          type: integer
          description: "Time spent in minutes"
        estimatedHours:
          type: number
          format: decimal
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        assigneeId:
          type: string
          format: uuid
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        estimatedHours:
          type: number
          format: decimal
        dueDate:
          type: string
          format: date-time
        labels:
          type: array
          items:
            type: string
    AssignTaskRequest:
      type: object
      required:
        - assigneeId
      properties:
        assigneeId:
          type: string
          format: uuid
    ChangeStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [TODO, IN_PROGRESS, IN_REVIEW, DONE, CANCELLED]
    TaskListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TaskDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer

    # Time Tracking schemas
...
    SearchResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    SearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [task, project, user, comment]
        title:
          type: string
        description:
          type: string
        projectName:
          type: string
        assigneeName:
          type: string
        createdAt:
          type: string
          format: date-time
        relevanceScore:
          type: number

    # Notification schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        status:
          type: integer
...