# server/src/main/java/com/taskboard/api/dto/AssignTaskRequest.java
package com.taskboard.api.dto;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * Request DTO for assigning a task to a user.
 * Contains the assignee ID for task assignment.
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AssignTaskRequest {

    @NotNull(message = "Assignee ID is required")
    private Long assigneeId;
}