# server/src/main/java/com/taskboard/api/dto/UpdateTaskRequest.java
package com.taskboard.api.dto;

import com.taskboard.api.enums.TaskPriority;
import com.taskboard.api.enums.TaskStatus;
import com.fasterxml.jackson.annotation.JsonFormat;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

/**
 * Request DTO for updating an existing task.
 * All fields are optional to allow partial updates.
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateTaskRequest {

    @Size(min = 1, max = 200, message = "Task title must be between 1 and 200 characters")
    private String title;

    @Size(max = 1000, message = "Task description cannot exceed 1000 characters")
    private String description;

    private TaskStatus status;

    private TaskPriority priority;

    private Long assigneeId;

    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
    private LocalDateTime dueDate;

    @Min(value = 0, message = "Estimated hours cannot be negative")
    @Max(value = 9999, message = "Estimated hours cannot exceed 9999")
    private Integer estimatedHours;

    @Min(value = 0, message = "Actual hours cannot be negative")
    private Integer actualHours;
}